Ankora & Kingsgrave

+++ src/MetalSlugTactics/MetalSlugTactics/Utils.cs
@@ -2,7 +2,7 @@
 
 public static class Utils
 {
-    internal static int FindLowestFrameRateMultipleAboveFifty(int originalRate)
+    internal static float FindLowestFrameRateMultipleAboveFifty(float originalRate)
     {
         for (var rate = originalRate / 2; rate > 50; rate--)
         {
